refactor: Split the codes
